Break down remaining actions
