+++ frontend/src/metabase/dashboard/actions/index.ts
@@ -1,8 +1,8 @@
-export * from "./actions";
 export * from "./core";
 export * from "./data-fetching";
 export * from "./navigation";
 export * from "./parameters";
+export * from "./revisions";
 export * from "./save";
 export * from "./sharing";
 export * from "./ui";
